requirements.txt
streamlit
scikit-learn
pandaspip install -r requirements.txt

matplotlib
plotly
